+++ src/configuration.js
@@ -1,6 +1,5 @@
 const cosmiconfig = require('cosmiconfig');
 import { readSync, readFileSync } from 'fs';
-import getGraphQLProjectConfig from 'graphql-config';
 import path from 'path';
 import { sync as globSync, hasMagic as globHasMagic } from 'glob';
 
@@ -17,9 +16,14 @@ export class Configuration {
       - schemaPaths: [string array] file(s) to read schema from
       - customRulePaths: [string array] path to additional custom rules to be loaded
       - stdin: [boolean] pass schema via stdin?
+      - commentDescriptions: [boolean] use old way of defining descriptions in GraphQL SDL
   */
   constructor(options = {}, stdinFd = null) {
-    const defaultOptions = { format: 'text', customRulePaths: [] };
+    const defaultOptions = {
+      format: 'text',
+      customRulePaths: [],
+      commentDescriptions: false,
+    };
     const configOptions = loadOptionsFromConfig(options.configDirectory);
 
     // TODO Get configs from .graphqlconfig file
@@ -34,20 +38,28 @@ export class Configuration {
     );
   }
 
+  getCommentDescriptions() {
+    return this.options.commentDescriptions;
+  }
+
   getSchema() {
     if (this.schema) {
       return this.schema;
     }
 
-    var schema;
-
     if (this.options.stdin) {
       this.schema = getSchemaFromFileDescriptor(this.stdinFd);
+      if (this.schema == null) {
+        return null;
+      }
+
       this.sourceMap = new SourceMap({ stdin: this.schema });
     } else if (this.options.schemaPaths) {
-      var expandedPaths = expandPaths(this.options.schemaPaths);
-      var segments = getSchemaSegmentsFromFiles(expandedPaths);
-
+      const expandedPaths = expandPaths(this.options.schemaPaths);
+      const segments = getSchemaSegmentsFromFiles(expandedPaths);
+      if (Object.keys(segments).length === 0) {
+        return null;
+      }
       this.sourceMap = new SourceMap(segments);
       this.schema = this.sourceMap.getCombinedSource();
     }
@@ -73,9 +85,8 @@ export class Configuration {
   }
 
   getRules() {
-    var rules = this.getAllRules();
-    var specifiedRules;
-
+    let rules = this.getAllRules();
+    let specifiedRules;
     if (this.options.rules && this.options.rules.length > 0) {
       specifiedRules = this.options.rules.map(toUpperCamelCase);
       rules = this.getAllRules().filter(rule => {
@@ -107,10 +118,11 @@ export class Configuration {
       return this.rules;
     }
 
-    var expandedPaths = expandPaths(this.rulePaths);
+    let expandedPaths = expandPaths(this.rulePaths);
     this.rules = [];
     expandedPaths.map(rulePath => {
-      var rule = Object.values(require(rulePath));
+      let ruleMap = require(rulePath);
+      let rule = Object.keys(ruleMap).map(k => ruleMap[k]);
 
       if (rule) {
         this.rules = this.rules.concat(rule);
@@ -124,7 +136,7 @@ export class Configuration {
     const issues = [];
 
     const ruleNames = this.getAllRules().map(rule => rule.name);
-    var misConfiguredRuleNames = []
+    let misConfiguredRuleNames = []
       .concat(
         this.options.only || [],
         this.options.except || [],
@@ -174,11 +186,24 @@ function loadOptionsFromConfig(configDirectory) {
 }
 
 function getSchemaFromFileDescriptor(fd) {
-  var b = new Buffer(1024);
-  var data = '';
+  let b = new Buffer(1024);
+  let data = '';
 
   while (true) {
-    var n = readSync(fd, b, 0, b.length);
+    let n;
+    try {
+      n = readSync(fd, b, 0, b.length);
+    } catch (e) {
+      if (e.code == 'EAGAIN') {
+        console.error(
+          'The --stdin option was specified, but not schema was provided via stdin.'
+        );
+      } else {
+        console.error(e.message);
+      }
+      return null;
+    }
+
     if (!n) {
       break;
     }
@@ -189,12 +214,20 @@ function getSchemaFromFileDescriptor(fd) {
 }
 
 function getSchemaFromFile(path) {
-  return readFileSync(path).toString('utf8');
+  try {
+    return readFileSync(path).toString('utf8');
+  } catch (e) {
+    console.error(e.message);
+  }
+  return null;
 }
 
 function getSchemaSegmentsFromFiles(paths) {
   return paths.reduce((segments, path) => {
-    segments[path] = getSchemaFromFile(path);
+    let schema = getSchemaFromFile(path);
+    if (schema) {
+      segments[path] = schema;
+    }
     return segments;
   }, {});
 }