+++ src/rules/enum_values_have_descriptions.js
@@ -1,11 +1,15 @@
 import { getDescription } from 'graphql/utilities/buildASTSchema';
 import { ValidationError } from '../validation_error';
 
-export function EnumValuesHaveDescriptions(context) {
+export function EnumValuesHaveDescriptions(configuration, context) {
   var isFixable = false;
   return {
     EnumValueDefinition(node, key, parent, path, ancestors) {
-      if (getDescription(node)) {
+      if (
+        getDescription(node, {
+          commentDescriptions: configuration.getCommentDescriptions(),
+        })
+      ) {
         return;
       }