+++ src/runner.js
@@ -28,6 +28,10 @@ export function run(stdout, stdin, stderr, argv) {
       '-p, --custom-rule-paths <paths>',
       'path to additional custom rules to be loaded. Example: rules/*.js'
     )
+    .option(
+      '--comment-descriptions',
+      'use old way of defining descriptions in GraphQL SDL'
+    )
     // DEPRECATED - This code should be removed in v1.0.0.
     .option(
       '-o, --only <rules>',
@@ -76,12 +80,15 @@ export function run(stdout, stdin, stderr, argv) {
   }
 
   const schema = configuration.getSchema();
+  if (schema == null) {
+    console.error('No valid schema input.');
+    return 2;
+  }
   const formatter = configuration.getFormatter();
   const rules = configuration.getRules();
   const schemaSourceMap = configuration.getSchemaSourceMap();
 
-  const errors = validateSchemaDefinition(schema, rules);
-
+  const errors = validateSchemaDefinition(schema, rules, configuration);
   const groupedErrors = groupErrorsBySchemaFilePath(errors, schemaSourceMap);
 
   stdout.write(formatter(groupedErrors));
@@ -131,6 +138,10 @@ function getOptionsFromCommander(commander) {
 
   if (commander.customRulePaths) {
     options.customRulePaths = commander.customRulePaths.split(',');
+  }
+
+  if (commander.commentDescriptions) {
+    options.commentDescriptions = commander.commentDescriptions;
   }
 
   if (commander.args && commander.args.length) {