+++ test/configuration.js
@@ -16,7 +16,6 @@ describe('Configuration', () => {
   author: User!
 }
 
-
 type Query {
   something: String!
 }
@@ -261,6 +260,18 @@ extend type Query {
       const issues = configuration.validate();
 
       assert.equal(issues.length, 2);
+    });
+  });
+
+  describe('getCommentDescriptions', () => {
+    it('defaults to false', () => {
+      const configuration = new Configuration({});
+      assert.equal(configuration.getCommentDescriptions(), false);
+    });
+
+    it('returns specified value', () => {
+      const configuration = new Configuration({ commentDescriptions: true });
+      assert.equal(configuration.getCommentDescriptions(), true);
     });
   });
 });